twister: runner: add method to check for changed dependencies
Using the ninja dependencies available after building and comparing them
with the changed source code relative to the merge base with the main,
the new method "dependencies_have_changed" checks if the current test
instance has any changed dependencies. In this fashion one does not
always have to run every test/sample project, but can skip those,
that are not affected by changes.

Signed-off-by: frei tycho <tfrei@baumer.com>